requirements.txt
requests==2.26.0
ftfy==6.0.3
jupyter==1.0.0
matplotlib==3.4.3
nltk==3.6.2
numpy==1.21.2
omegaconf==2.1.1
opencv-python==4.6.0
pandas==1.3.3
Pillow==8.3.2
pycocotools==2.0.2
pytest==6.2.5
regex==2021.8.28
scikit_learn==0.24.2
scipy==1.7.1
scikit-image==0.18.3
torch==1.9.0
torchray==1.0.0.2
torchvision==0.10.0
tqdm==4.62.2
wandb
